refactor(bank-sync,app): consolidate Erste parsing into singleton classes
- Drop "Modern" qualifier (no other Erste parser variant exists):
  ErsteModernPositionalParser → ErsteParser, ERSTE_MODERN_PAGE_NOISE_PATTERNS → ERSTE_PAGE_NOISE_PATTERNS, file → erste.parser.ts.

- Bundle account-info finders (find-iban / -statement-date /
  -old-balance / -new-balance + orchestrator util) into
  ErsteAccountInfoExtractor singleton with @Log on extract().

- Bundle row-grouping (group + comparator) into ErsteRowGrouper
  singleton with @Log on group(). RowBucket renamed → ErsteRowBucket
  in its own file.

- Convert parse-erste-card-merchant.util into ErsteCardMerchantParser
  singleton with @Log on parse(). Same for
  generate-erste-external-id.util → ErsteExternalIdGenerator with @Log
  on generate() (helps debug external-id collisions).

- Prefix all Erste-specific interfaces with Erste:
  PageRowInterface, DateAmountInterface, InlineDateAmountInterface,
  DateAmountInputInterface. PdfTextItemInterface stays bank-agnostic
  (raw native-module shape). Parser-state and transaction-accumulator
  classes prefixed similarly.

- Delete redundant parse-erste-items.util.ts wrapper; ErsteFileClient
  now calls ersteParser singleton directly.

- ErsteSyncService.parseFile drops dead const-capture pattern (the
  closures already work against the local ersteClient instance).

- use-quick-import.hook drops ad-hoc getLogger calls; @Log decorators
  on the underlying services already cover lifecycle.
