+++ packages/app/src/sync/hook/use-quick-import.hook.ts
@@ -1,12 +1,9 @@
-import { getLogger } from '@budgie/logger';
 import * as DocumentPicker from 'expo-document-picker';
 import { useState } from 'react';
 
 import { getErrorMessage, isDefined, isNotEmptyString } from '@rnw-community/shared';
 
 import { QuickImportConfigInterface } from '../interface/quick-import-config.interface';
-
-const logger = getLogger('useQuickImport');
 
 interface QuickImportResult {
     readonly isLoading: boolean;
@@ -21,35 +18,27 @@ export const useQuickImport = (config: QuickImportConfigInterface | null): Quick
 
     const handleQuickImport = () => {
         if (!isDefined(config)) {
-            logger.log('handleQuickImport:skipped:no-config');
-
             return;
         }
 
         const execute = async (): Promise<void> => {
-            logger.log('picker:start', { mimeType: config.mimeType });
             setIsLoading(true);
             setError(null);
 
             const result = await DocumentPicker.getDocumentAsync({ type: config.mimeType, copyToCacheDirectory: true });
-            const { uri, name, size } = result.assets?.at(0) ?? {};
-            logger.log('picker:done', { canceled: result.canceled, hasUri: isNotEmptyString(uri), name: name ?? '', sizeBytes: size ?? 0 });
+            const uri = result.assets?.at(0)?.uri;
 
             if (result.canceled || !isNotEmptyString(uri)) {
                 setIsLoading(false);
-                logger.log('picker:cancelled-or-empty');
-
-                return;
+                
+return;
             }
 
-            logger.log('importHandler:start', { uri });
             await config.importHandler(uri);
-            logger.log('importHandler:done', { uri });
         };
 
         void execute()
             .catch((importError: unknown) => {
-                logger.error('execute:failed', { errorMessage: getErrorMessage(importError) });
                 setError(getErrorMessage(importError));
             })
             .finally(() => {
