+++ packages/bank-sync/src/erste/client/erste-file.client.ts
@@ -2,7 +2,7 @@ import { isDefined } from '@rnw-community/shared';
 
 import { ersteAccountMapper } from '../mapper/erste-account.mapper';
 import { ersteTransactionMapper } from '../mapper/erste-transaction.mapper';
-import { parseErsteItems } from '../util/parse-erste-items.util';
+import { ersteParser } from '../parser/erste.parser';
 
 import type { BankAccountInterface } from '../../core/interface/bank-account.interface';
 import type { BankTransactionInterface } from '../../core/interface/bank-transaction.interface';
@@ -14,7 +14,7 @@ export class ErsteFileClient {
     private parsedData: ErsteParsedDataInterface | null = null;
 
     parse(items: PdfTextItemInterface[]): void {
-        this.parsedData = parseErsteItems(items);
+        this.parsedData = ersteParser.parse(items);
     }
 
     getAccounts(): BankAccountInterface[] {
