+++ packages/bank-sync/src/erste/interface/erste-date-amount-input.interface.ts
@@ -1,4 +1,4 @@
-export interface DateAmountInputInterface {
+export interface ErsteDateAmountInputInterface {
     readonly day: string;
     readonly month: string;
     readonly year: string;
