+++ packages/bank-sync/src/erste/interface/erste-date-amount.interface.ts
@@ -1,4 +1,4 @@
-export interface DateAmountInterface {
+export interface ErsteDateAmountInterface {
     readonly date: Date;
     readonly amount: number;
     readonly isCredit: boolean;
