+++ packages/bank-sync/src/erste/interface/erste-inline-date-amount.interface.ts
@@ -1,0 +1,5 @@
+import type { ErsteDateAmountInterface } from './erste-date-amount.interface';
+
+export interface ErsteInlineDateAmountInterface extends ErsteDateAmountInterface {
+    readonly prefix: string;
+}
