+++ packages/bank-sync/src/erste/interface/erste-page-row.interface.ts
@@ -1,6 +1,6 @@
 import type { PdfTextItemInterface } from './pdf-text-item.interface';
 
-export interface PageRowInterface {
+export interface ErstePageRowInterface {
     readonly page: number;
     readonly y: number;
     readonly leftItems: PdfTextItemInterface[];
