--- packages/bank-sync/src/erste/interface/inline-date-amount.interface.ts
@@ -1,5 +1,0 @@
-import type { DateAmountInterface } from './date-amount.interface';
-
-export interface InlineDateAmountInterface extends DateAmountInterface {
-    readonly prefix: string;
-}
